backend/requirements.txt
flask-admin==1.6.1
flask==3.0.0
flask-cors==4.0.0
flask-login==0.6.3
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
google-generativeai==0.3.1
deep-translator==1.11.4
requests>=2.31.0
Pillow>=10.2.0
